+++ extension.js
@@ -1,36 +1,100 @@
-// The module 'vscode' contains the VS Code extensibility API
-// Import the module and reference it with the alias vscode in your code below
 const vscode = require('vscode');
+const fs = require('fs');
+const path = require('path');
 
-// This method is called when your extension is activated
-// Your extension is activated the very first time the command is executed
+let scriptTerminal = null;
+
+class ScriptTreeItem extends vscode.TreeItem {
+    constructor(scriptName, scriptCmd) {
+        super(scriptName, vscode.TreeItemCollapsibleState.None);
+        this.description = scriptCmd;
+        this.contextValue = 'scriptItem';
+        this.iconPath = new vscode.ThemeIcon('wrench'); 
+    }
+}
+
+class ScriptsProvider {
+    constructor(workspaceRoot) {
+        this.workspaceRoot = workspaceRoot;
+        this._onDidChangeTreeData = new vscode.EventEmitter();
+        this.onDidChangeTreeData = this._onDidChangeTreeData.event;
+    }
+
+    refresh() {
+        this._onDidChangeTreeData.fire();
+    }
+
+    getTreeItem(element) {
+        return element;
+    }
+
+    getChildren() {
+        const packageJsonPath = path.join(this.workspaceRoot, 'package.json');
+        if (fs.existsSync(packageJsonPath)) {
+            const pkg = JSON.parse(fs.readFileSync(packageJsonPath, 'utf8'));
+            const scripts = pkg.scripts || {};
+            return Object.keys(scripts).map(scriptName => {
+                return new ScriptTreeItem(scriptName, scripts[scriptName]);
+            });
+        }
+        return [];
+    }
+}
 
 /**
  * @param {vscode.ExtensionContext} context
  */
 function activate(context) {
+    const workspaceFolders = vscode.workspace.workspaceFolders;
+    if (!workspaceFolders) {
+        return;
+    }
+    const workspaceRoot = workspaceFolders[0].uri.fsPath;
+    const packageJsonPath = path.join(workspaceRoot, 'package.json');
 
-	// Use the console to output diagnostic information (console.log) and errors (console.error)
-	// This line of code will only be executed once when your extension is activated
-	console.log('Congratulations, your extension "run-project" is now active!');
+    if (!fs.existsSync(packageJsonPath)) {
+        vscode.window.showWarningMessage('No package.json found in workspace.');
+        return;
+    }
 
-	// The command has been defined in the package.json file
-	// Now provide the implementation of the command with  registerCommand
-	// The commandId parameter must match the command field in package.json
-	const disposable = vscode.commands.registerCommand('run-project.helloWorld', function () {
-		// The code you place here will be executed every time your command is executed
+    const scriptsProvider = new ScriptsProvider(workspaceRoot);
+    vscode.window.registerTreeDataProvider('run-project.scriptsView', scriptsProvider);
 
-		// Display a message box to the user
-		vscode.window.showInformationMessage('Hello World from ScriptDeck!');
-	});
+    context.subscriptions.push(
+        vscode.commands.registerCommand('run-project.refreshScripts', () => scriptsProvider.refresh())
+    );
 
-	context.subscriptions.push(disposable);
+    context.subscriptions.push(
+        vscode.commands.registerCommand('run-project.runScript', (item) => {
+            if (!item || !item.label) return;
+            if (scriptTerminal) {
+                scriptTerminal.dispose();
+            }
+            scriptTerminal = vscode.window.createTerminal(`Script: ${item.label}`);
+            scriptTerminal.show();
+            scriptTerminal.sendText(`npm run ${item.label}`);
+        })
+    );
+
+    context.subscriptions.push(
+        vscode.commands.registerCommand('run-project.stopScript', () => {
+            if (scriptTerminal) {
+                scriptTerminal.dispose();
+                scriptTerminal = null;
+            }
+        })
+    );
 }
 
-// This method is called when your extension is deactivated
-function deactivate() {}
+function deactivate() {
+    if (scriptTerminal) {
+        scriptTerminal.dispose();
+    }
+}
 
 module.exports = {
-	activate,
-	deactivate
-}
+    activate,
+    deactivate,
+    ScriptTreeItem,
+    ScriptsProvider
+};